src/app/homealumno/homealumno.page.spec.ts
import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
import { IonicModule } from '@ionic/angular';
import { HomealumnoPage } from './homealumno.page';

describe('HomealumnoPage', () => {
  let component: HomealumnoPage;
  let fixture: ComponentFixture<HomealumnoPage>;

  beforeEach(waitForAsync(() => {
    TestBed.configureTestingModule({
      declarations: [HomealumnoPage],
      imports: [IonicModule.forRoot()]
    }).compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(HomealumnoPage);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should have a defined qrData property', () => {
    expect(component.qrData).toBeDefined();
  });

  it('should have a defined scanQRCode method', () => {
    expect(component.scanQRCode).toBeDefined();
  });

  it('should have a defined decodeQRCode method', () => {
    expect(component.decodeQRCode).toBeDefined();
  });

  it('should have a defined decodeDataUrl method', () => {
    expect(component.decodeDataUrl).toBeDefined();
  });

  it('should have a defined sendEmail method', () => {
    expect(component.sendEmail).toBeDefined();
  });

  // Agrega más pruebas según sea necesario para cubrir tus funciones y casos de uso.
});
